+++ packages/remix-architect/__tests__/server-test.ts
@@ -90,7 +90,10 @@ describe("architect createRequestHandler", () => {
         return new Response(`URL: ${new URL(req.url).pathname}`);
       });
 
-      await lambdaTester(createRequestHandler({ build: undefined } as any))
+      // We don't have a real app to test, but it doesn't matter. We won't ever
+      // call through to the real createRequestHandler
+      // @ts-expect-error
+      await lambdaTester(createRequestHandler({ build: undefined }))
         .event(createMockEvent({ rawPath: "/foo/bar" }))
         .expectResolve((res) => {
           expect(res.statusCode).toBe(200);
@@ -103,7 +106,10 @@ describe("architect createRequestHandler", () => {
         return new Response(null, { status: 200 });
       });
 
-      await lambdaTester(createRequestHandler({ build: undefined } as any))
+      // We don't have a real app to test, but it doesn't matter. We won't ever
+      // call through to the real createRequestHandler
+      // @ts-expect-error
+      await lambdaTester(createRequestHandler({ build: undefined }))
         .event(createMockEvent({ rawPath: "/foo/bar" }))
         .expectResolve((res) => {
           expect(res.statusCode).toBe(200);
@@ -115,7 +121,10 @@ describe("architect createRequestHandler", () => {
         return new Response(null, { status: 204 });
       });
 
-      await lambdaTester(createRequestHandler({ build: undefined } as any))
+      // We don't have a real app to test, but it doesn't matter. We won't ever
+      // call through to the real createRequestHandler
+      // @ts-expect-error
+      await lambdaTester(createRequestHandler({ build: undefined }))
         .event(createMockEvent({ rawPath: "/foo/bar" }))
         .expectResolve((res) => {
           expect(res.statusCode).toBe(204);
@@ -142,7 +151,10 @@ describe("architect createRequestHandler", () => {
         return new Response(null, { headers });
       });
 
-      await lambdaTester(createRequestHandler({ build: undefined } as any))
+      // We don't have a real app to test, but it doesn't matter. We won't ever
+      // call through to the real createRequestHandler
+      // @ts-expect-error
+      await lambdaTester(createRequestHandler({ build: undefined }))
         .event(createMockEvent({ rawPath: "/" }))
         .expectResolve((res) => {
           expect(res.statusCode).toBe(200);